requirements.txt
tensorflow==1.15.0
keras==2.3.1
numpy=1.19.2
pillow==8.1.0